requirements.txt
asgiref==3.2.10
attrs==20.3.0
bleach==3.1.4
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
dj-database-url==0.3.0
dj-static==0.0.6
Django==3.1.4
django-braces==1.14.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.0
django-jet==1.0.8
django-oauth-toolkit==1.3.3
django-rest-framework-social-oauth2==1.1.0
django-storages==1.10.1
Django-Verify-Email==0.0.3
djangorestframework==3.12.2
dropbox==11.0.0
gunicorn==20.0.4
idna==2.10
Markdown==2.6.6
more-itertools==8.6.0
numpy==1.19.4
oauthlib==3.1.0
packaging==20.7
pandas==1.1.4
Pillow==7.1.0
pluggy==0.13.1
ply==3.11
py==1.9.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
python-dateutil==2.8.1
python-decouple==3.0
python-dotenv==0.12.0
python3-openid==3.2.0
pytz==2020.1
razorpay==1.2.0
requests==2.25.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
static3==0.7.0
stone==3.2.1
Unipath==1.0
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.0.1
